Moving Charts generation to its very own service
(cherry picked from commit 9f6f9f268b54775ce2180afd2ce90012837f4f67)
